+++ README.md
@@ -22,7 +22,7 @@ Here's a small map of the project:
   Declarative definition of the database tables we use. These include parts of the `garfield` database (into which we save all documents and their associated data) and parts of the `fsmi` database (which we query for user auth)
 * `accounting.py`  
   The accounting backend. We insert all accounting information into the `garfield` database through this. Should garfield itself ever die and get replaced with something else, this is where you'll need to add new accounting hooks.
-* `apigen.py`  
-  We're using Flask, Flask-login, SQLAlchemy, jsonquery and marshmallow. That means a lot of work is done for us and what's left of most API endpoints is mostly boilerplate for stringing all of this together. `apigen.py` is where we abstract all of that away, too, so most API endpoints can be generated with a simple call to `apigen.endpoint`.
+* `api_utils.py`  
+  We're using Flask, Flask-login, SQLAlchemy, jsonquery and marshmallow. That means a lot of work is done for us and what's left of most API endpoints is mostly boilerplate for stringing all of this together. `api_utils.py` is where we abstract all of that away, too, so most API endpoints can be generated with a simple call to `api_utils.endpoint`. Also found there: grab bag of small utility functions that would only clutter routes.py.
 
 The rest of the files should be fairly straight-forward. Godspeed.
