+++ routes.py
@@ -1,7 +1,7 @@
 #! /usr/bin/env python3
 
 import accounting
-import apigen
+import api_utils
 import config
 import serialization_schemas as schemas
 
@@ -9,7 +9,7 @@ from flask import request
 from flask.ext.login import login_user, logout_user, current_user, login_required
 
 from odie import app, db, ClientError
-from apigen import endpoint, filtered_results, api_route
+from api_utils import endpoint, filtered_results, api_route
 from models.documents import Lecture, Deposit, Document, Examinant
 from models.odie import Order
 from models.public import User
@@ -48,7 +48,7 @@ def _lectures(document_ids):
 
 
 @api_route('/api/print', methods=['POST'])
-@apigen.deserialize(schemas.PrintJobLoadSchema)
+@api_utils.deserialize(schemas.PrintJobLoadSchema)
 @login_required
 def print_documents(data):
     documents = [Document.query.get(id) for id in data['document_ids']]  # TODO sort documents in python
@@ -84,7 +84,7 @@ def print_documents(data):
 
 @api_route('/api/log_erroneous_sale', methods=['POST'])
 @login_required
-@apigen.deserialize(schemas.ErroneousSaleLoadSchema)
+@api_utils.deserialize(schemas.ErroneousSaleLoadSchema)
 def accept_erroneous_sale(data):
     accounting.log_erroneous_sale(data['amount'], current_user, data['cash_box'])
     db.session.commit()
@@ -93,7 +93,7 @@ def accept_erroneous_sale(data):
 
 @api_route('/api/log_deposit_return', methods=['POST'])
 @login_required
-@apigen.deserialize(schemas.DepositLoadSchema)
+@api_utils.deserialize(schemas.DepositLoadSchema)
 def log_deposit_return(data):
     dep = Deposit.query.get(data['id'])
     db.session.delete(dep)
@@ -104,7 +104,7 @@ def log_deposit_return(data):
 
 @api_route('/api/donation', methods=['POST'])
 @login_required
-@apigen.deserialize(schemas.DonationLoadSchema)
+@api_utils.deserialize(schemas.DonationLoadSchema)
 def log_donation(data):
     accounting.log_donation(data['amount'], data['cash_box'])
     db.session.commit()
@@ -113,7 +113,7 @@ def log_donation(data):
 
 api_route('/api/orders', methods=['GET'])(
 login_required(
-apigen.endpoint(
+api_utils.endpoint(
         schemas={
             'GET': schemas.OrderDumpSchema,
             'POST': schemas.OrderLoadSchema
@@ -122,7 +122,7 @@ apigen.endpoint(
 ))
 
 api_route('/api/orders', methods=['POST'])(
-apigen.endpoint(
+api_utils.endpoint(
         schemas={
             'POST': schemas.OrderLoadSchema
         },
@@ -131,7 +131,7 @@ apigen.endpoint(
 
 api_route('/api/orders/<int:instance_id>', methods=['GET', 'DELETE'])(
 login_required(
-apigen.endpoint(
+api_utils.endpoint(
         schemas={'GET': schemas.OrderDumpSchema},
         query=Order.query,
         allow_delete=True)
@@ -139,7 +139,7 @@ apigen.endpoint(
 
 
 api_route('/api/lectures')(
-apigen.endpoint(
+api_utils.endpoint(
         schemas={'GET': schemas.LectureSchema},
         query=Lecture.query)
 )
@@ -147,11 +147,11 @@ apigen.endpoint(
 @api_route('/api/lectures/<int:id>/documents')
 def lecture_documents(id):
     lecture = Lecture.query.get(id)
-    return apigen.filtered_results(lecture.documents, schemas.DocumentSchema)
+    return api_utils.filtered_results(lecture.documents, schemas.DocumentSchema)
 
 
 api_route('/api/examinants')(
-apigen.endpoint(
+api_utils.endpoint(
         schemas={'GET': schemas.ExaminantSchema},
         query=Examinant.query)
 )
@@ -159,17 +159,17 @@ apigen.endpoint(
 @api_route('/api/examinants/<int:id>/documents')
 def examinant_documents(id):
     examinant = Examinant.query.get(id)
-    return apigen.filtered_results(examinant.documents, schemas.DocumentSchema)
+    return api_utils.filtered_results(examinant.documents, schemas.DocumentSchema)
 
 api_route('/api/documents')(
-apigen.endpoint(
+api_utils.endpoint(
         schemas={'GET': schemas.DocumentSchema},
         query=Document.query)
 )
 
 api_route('/api/deposits')(
 login_required(
-apigen.endpoint(
+api_utils.endpoint(
         schemas={'GET': schemas.DepositDumpSchema},
         query=Deposit.query)
 ))
